fix: :truck: Moved block components to utils
Issue: #39
