+++ src/block/spawn.rs
@@ -5,90 +5,8 @@ use crate::{
     AppState,
 };
 use crate::utils::blockdata::*;
-use super::{
-    BLOCK_SIZE,
-    BLOCK_POSITION,
-    SpawnEvent,
-    CurrentBlock,
-    PlayerBlock,
-    Block,
-    BlockType,
-};
-
-impl BlockType {
-    fn color(&self) -> Color {
-        match self {
-            BlockType::TypeI => I_COLOR,
-            BlockType::TypeJ => J_COLOR,
-            BlockType::TypeL => L_COLOR,
-            BlockType::TypeO => O_COLOR,
-            BlockType::TypeS => S_COLOR,
-            BlockType::TypeT => T_COLOR,
-            BlockType::TypeZ => Z_COLOR,
-        }
-    }
-    // このメソッドではブロックを生成する時のポジション(x,y)を返します。
-    // ブロックデータは配列x4のusizex16からなり、usizex16から縦4マス横4マスの
-    // ブロックの範囲とします。
-    // そこからメソッドに渡されたIDを元にブロックデータの値を参照し、
-    // 一致した箇所をxy軸に変換して返します。
-    // 例
-    // BlockType::TypeI.position(2) -> INITIAL_POSITION + Vec2::new(GRID_SIZE * 1, -GRID_SIZE * 1)
-    pub fn position(
-        &self,
-        init_pos: Vec2,
-        blockdata_id: usize,
-        block_id: usize,
-    ) -> Vec2 {
-        let closure = |block: [[usize; 16]; 4]| {
-            let mut position = init_pos;
-
-            for i in 0..block[blockdata_id].len() {
-                if block_id == block[blockdata_id][i] {
-                    // trace!("position: {}", position);
-                    return position
-                }
-                position.x += GRID_SIZE;
-                // movement y
-                if i % 4 == 3 {
-                    position.x = init_pos.x;
-                    position.y -= GRID_SIZE;
-                }
-            }
-            // id should be in the block[0]
-            panic!("id: {} is not found", block_id);
-        };
-
-        match self {
-            BlockType::TypeI => closure(I_BLOCK),
-            BlockType::TypeJ => closure(J_BLOCK),
-            BlockType::TypeL => closure(L_BLOCK),
-            BlockType::TypeO => closure(O_BLOCK),
-            BlockType::TypeS => closure(S_BLOCK),
-            BlockType::TypeT => closure(T_BLOCK),
-            BlockType::TypeZ => closure(Z_BLOCK),
-        }
-    }
-}
-
-impl Block {
-    fn new(
-        blockdata_id: usize,
-        block_id: usize,
-        block: BlockType,
-    ) -> (Self, Sprite, Transform, PlayerBlock) {
-        (
-            Self,
-            Sprite::from_color(block.color(), Vec2::ONE),
-            Transform {
-                translation: block.position(BLOCK_POSITION, blockdata_id, block_id).extend(1.0),
-                scale: BLOCK_SIZE.extend(1.0),
-                ..Default::default()
-            },
-            PlayerBlock(block_id),
-        )
-    }
-}
+use super::SpawnEvent;
+use crate::utils::prelude::*;
 
 fn setup(
     mut events: EventWriter<SpawnEvent>,
