+++ src/utils/mod.rs
@@ -1,12 +1,15 @@
 use bevy::prelude::*;
 
+pub mod block;
 pub mod blockdata;
+pub mod prelude;
 
 pub struct UtilsPlugin;
 
 impl Plugin for UtilsPlugin {
     fn build(&self, app: &mut App) {
         app
+            .add_plugins(block::BlockPlugin)
         ;
     }
 }
