+++ src/utils/prelude.rs
@@ -1,0 +1,6 @@
+pub use crate::utils::block::{
+    BLOCK_SPEED,
+    CurrentBlock,
+    PlayerBlock,
+    Block,
+};
